+++ bin/make_cov.py
@@ -1,6 +1,6 @@
 import matplotlib
 matplotlib.use('Agg')
-from tilec import kspace,ilc
+from tilec import kspace,ilc,pipeline
 import numpy as np
 import os,sys
 from pixell import enmap
@@ -26,7 +26,6 @@ parser.add_argument("version", type=str,help='Version name.')
 parser.add_argument("region", type=str,help='Region name.')
 parser.add_argument("arrays", type=str,help='Comma separated list of array names. Array names map to a data specification in data.yml')
 parser.add_argument("--mask-version", type=str,  default="padded_v1",help='Mask version')
-parser.add_argument("--nsim",     type=int,  default=None,help="Number of sims. If not specified, runs on data.")
 parser.add_argument("-o", "--overwrite", action='store_true',help='Ignore existing version directory.')
 parser.add_argument("-m", "--memory-intensive", action='store_true',help='Do not save FFTs to scratch disk. Can be faster, but very memory intensive.')
 parser.add_argument("--uncalibrated", action='store_true',help='Do not use calibration factors.')
@@ -39,105 +38,8 @@ parser.add_argument("--rfit-lmin",     type=int,  default=pdefaults['rfit_lmin']
 
 args = parser.parse_args()
 
-save_scratch = not(args.memory_intensive)
-save_path = sints.dconfig['tilec']['save_path']
-scratch_path = sints.dconfig['tilec']['scratch_path']
-savedir = save_path + args.version + "/" + args.region +"/"
-if save_scratch: scratch = scratch_path + args.version + "/" + args.region +"/"
-if not(args.overwrite):
-    assert not(os.path.exists(savedir)), \
-   "This version already exists on disk. Please use a different version identifier."
-try: os.makedirs(savedir)
-except:
-    if args.overwrite: pass
-    else: raise
-if save_scratch:     
-    try: os.makedirs(scratch)
-    except: pass
-gconfig = io.config_from_yaml("input/data.yml")
-
-
-mask = sints.get_act_mr3_crosslinked_mask(args.region,
-                                          version=args.mask_version,
-                                          kind='binary_apod')
-shape,wcs = mask.shape,mask.wcs
-
-with bench.show("ffts"):
-    kcoadds = []
-    ksplits = []
-    wins = []
-    lmins = []
-    lmaxs = []
-    hybrids = []
-    do_radial_fit = []
-    save_names = [] # to make sure nothing is overwritten
-    friends = {} # what arrays are each correlated with?
-    names = []
-    for array in args.arrays.split(','):
-        ainfo = gconfig[array]
-        dm = sints.models[ainfo['data_model']](region=mask,calibrated=not(args.uncalibrated))
-        name = ainfo['id']
-        names.append(name)
-        rfit = ainfo['radial_fit']
-        assert isinstance(rfit,bool)
-        do_radial_fit.append(rfit)
-        try: friends[name] = ainfo['correlated']
-        except: friends[name] = None
-        hybrids.append(ainfo['hybrid_average'])
-        ksplit,kcoadd,win = kspace.process(dm,args.region,name,mask,ncomp=1,skip_splits=False)
-        if save_scratch: 
-            kcoadd_name = savedir + "kcoadd_%s.hdf" % array
-            ksplit_name = scratch + "ksplit_%s.hdf" % array
-            win_name = scratch + "win_%s.hdf" % array
-            assert win_name not in save_names
-            assert kcoadd_name not in save_names
-            assert ksplit_name not in save_names
-            enmap.write_map(win_name,win)
-            enmap.write_map(kcoadd_name,kcoadd)
-            enmap.write_map(ksplit_name,ksplit)
-            wins.append(win_name)
-            kcoadds.append(kcoadd_name)
-            ksplits.append(ksplit_name)
-            save_names.append(win_name)
-            save_names.append(kcoadd_name)
-            save_names.append(ksplit_name)
-        else:
-            wins.append(win.copy())
-            kcoadds.append(kcoadd.copy())
-            ksplits.append(ksplit.copy())
-        lmins.append(ainfo['lmin'])
-        lmaxs.append(ainfo['lmax'])
-
-# Decide what pairs to do hybrid smoothing for
-narrays = len(args.arrays.split(','))
-anisotropic_pairs = []
-for i in range(narrays):
-    for j in range(i,narrays):
-        name1 = names[i]
-        name2 = names[j]
-        if (i==j) and (hybrids[i] and hybrids[j]):
-            anisotropic_pairs.append((i,j))
-            continue
-        if (friends[name1] is None) or (friends[name2] is None): continue
-        if name2 in friends[name1]:
-            assert name1 in friends[name2], "Correlated arrays spec is not consistent."
-            anisotropic_pairs.append((i,j))
-        
-print("Anisotropic pairs: ",anisotropic_pairs)
-
-enmap.write_map(savedir+"tilec_mask.fits",mask)
-save_fn = lambda x,a1,a2: enmap.write_map(savedir+"tilec_hybrid_covariance_%s_%s.hdf" % (names[a1],names[a2]),enmap.enmap(x,wcs))
-
-maxval = ilc.build_empirical_cov(ksplits,kcoadds,wins,mask,lmins,lmaxs,
-                        anisotropic_pairs,do_radial_fit,save_fn,
-                        signal_bin_width=args.signal_bin_width,
-                        signal_interp_order=args.signal_interp_order,
-                        dfact=(args.dfact,args.dfact),
-                        rfit_lmaxes=None,
-                        rfit_wnoise_width=args.rfit_wnoise_width,
-                        rfit_lmin=args.rfit_lmin,
-                        rfit_bin_width=None,
-                        verbose=True,
-                        debug_plots_loc=savedir)
-
-np.savetxt(savedir + "maximum_value_of_covariances.txt",np.array([[maxval]]))
+pipeline.build_and_save_cov(args.arrays,args.region,args.version,args.mask_version,
+                       args.signal_bin_width,args.signal_interp_order,args.dfact,
+                       args.rfit_wnoise_width,args.rfit_lmin,
+                       args.overwrite,args.memory_intensive,args.uncalibrated,
+                       sim_splits=None)
